requirements.txt
python-dotenv==0.19.0
Django==3.2.6
psycopg2==2.9.1
django-modeltranslation==0.17.3
pandas==1.3.1
demjson==2.2.4
SQLAlchemy==1.4.22